refac(rfr): move `Callsite` and `CallsiteId` out of `chunked` module
The callsite and its Id were originally created in the chunked module,
even though they are used in both chunked and streamed recording
formats. This is a bit confusing.

This change moves `Callsite` and `CallsiteId` to a new `callsite` module
in the root of the `rfr` crate and then exposes them at the root of the
crate (similar to what we do for all other public types that aren't in
the `chunked` or `streamed` modules.

Additionally, a few places where common types were still being referred
to within the `rfr` crate via the `common` module have been modified to
refer to the types at the root of the crate.
